refactor: 게임로직들을 포함한 폴더를 lotto가 아닌 domain으로 변경
