+++ src/domain/LottoGame.js
